+++ tests/Upwork/API/Routers/AuthTest.php
@@ -8,7 +8,7 @@ class AuthTest extends CommonTestRouter
     /**
      * Setup
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
     }