+++ tests/Upwork/API/Routers/Hr/InterviewsTest.php
@@ -10,7 +10,7 @@ class InterviewsTest extends CommonTestRouter
     /**
      * Setup
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
     }