+++ tests/Upwork/API/Routers/MetadataTest.php
@@ -8,7 +8,7 @@ class MetadataTest extends CommonTestRouter
     /**
      * Setup
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
     }