+++ tests/Upwork/API/Routers/Organization/UsersTest.php
@@ -10,7 +10,7 @@ class UsersTest extends CommonTestRouter
     /**
      * Setup
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
     }