+++ tests/Upwork/API/Routers/Reports/Finance/BillingsTest.php
@@ -10,7 +10,7 @@ class BillingsTest extends CommonTestRouter
     /**
      * Setup
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
     }