+++ tests/Upwork/API/Routers/SnapshotTest.php
@@ -8,7 +8,7 @@ class SnapshotTest extends CommonTestRouter
     /**
      * Setup
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
     }