+++ tests/Upwork/API/UtilsTest.php
@@ -1,11 +1,12 @@
 <?php
 require __DIR__ . '/../../../vendor/autoload.php';
 
+use PHPUnit\Framework\TestCase;
 use Upwork\API\Config as Apiconfig;
 use Upwork\API\Debug as ApiDebug;
 use Upwork\API\Utils as Utils;
 
-class UtilsTest extends \PHPUnit_Framework_TestCase
+class UtilsTest extends TestCase
 {
     /**
      * @test