+++ CHANGES.md
@@ -1,7 +1,19 @@
 # Release History
 
 ## 1.4.0
+* Send Message to a Batch of Rooms API
+* Workdiary::get was replaced with Workdiary::getByCompany (breaking!)
+* Milestones::delete was replaced with Milestones::deleteMilestone (breaking!)
+* Migrate tests
+* Bug fixes
 * Stop supporting some Reports API according to announce from Thursday, 2018-06-28
+
+## 1.3.3
+* Add Room Messages API
+
+## 1.3.2
+* Add Specialties API
+* Add Skills V2 API
 
 ## 1.3.1
 * Fixed non-working getByContract from Workdays API